# Remove fenwick tree