… internal/fenwick/LICENSE
… internal/fenwick/list.go