First commit adds a threshold implemented in `useFenwickTree` that disables fenwick tree until some condition is met. Following is fenwick_disabled vs fenwick_threshold:

```
benchmark                                              old ns/op     new ns/op     delta
BenchmarkTDigestAddOnce/compression=1-4                165           156           -5.45%
BenchmarkTDigestAddOnce/compression=10-4               203           210           +3.45%
BenchmarkTDigestAddOnce/compression=20-4               219           218           -0.46%
BenchmarkTDigestAddOnce/compression=30-4               231           229           -0.87%
BenchmarkTDigestAddOnce/compression=50-4               243           241           -0.82%
BenchmarkTDigestAddOnce/compression=100-4              259           260           +0.39%
BenchmarkTDigestAddMulti/compression=1_n=10-4          1905          1561          -18.06%
BenchmarkTDigestAddMulti/compression=1_n=100-4         16277         14633         -10.10%
BenchmarkTDigestAddMulti/compression=1_n=1000-4        177664        165089        -7.08%
BenchmarkTDigestAddMulti/compression=1_n=10000-4       1616827       1536685       -4.96%
BenchmarkTDigestAddMulti/compression=10_n=10-4         3172          2228          -29.76%
BenchmarkTDigestAddMulti/compression=10_n=100-4        24701         16744         -32.21%
BenchmarkTDigestAddMulti/compression=10_n=1000-4       238823        200966        -15.85%
BenchmarkTDigestAddMulti/compression=10_n=10000-4      2154064       2062918       -4.23%
BenchmarkTDigestAddMulti/compression=20_n=10-4         4229          2646          -37.43%
BenchmarkTDigestAddMulti/compression=20_n=100-4        36674         20279         -44.70%
BenchmarkTDigestAddMulti/compression=20_n=1000-4       280544        224492        -19.98%
BenchmarkTDigestAddMulti/compression=20_n=10000-4      2417438       2288713       -5.32%
BenchmarkTDigestAddMulti/compression=30_n=10-4         5160          3059          -40.72%
BenchmarkTDigestAddMulti/compression=30_n=100-4        52129         22113         -57.58%
BenchmarkTDigestAddMulti/compression=30_n=1000-4       387922        245660        -36.67%
BenchmarkTDigestAddMulti/compression=30_n=10000-4      2607715       2420602       -7.18%
BenchmarkTDigestAddMulti/compression=50_n=10-4         6190          3606          -41.74%
BenchmarkTDigestAddMulti/compression=50_n=100-4        65618         25500         -61.14%
BenchmarkTDigestAddMulti/compression=50_n=1000-4       533828        272792        -48.90%
BenchmarkTDigestAddMulti/compression=50_n=10000-4      3028039       2732537       -9.76%
BenchmarkTDigestAddMulti/compression=100_n=10-4        9506          4858          -48.90%
BenchmarkTDigestAddMulti/compression=100_n=100-4       79124         31295         -60.45%
BenchmarkTDigestAddMulti/compression=100_n=1000-4      1108901       347337        -68.68%
BenchmarkTDigestAddMulti/compression=100_n=10000-4     4894191       3341165       -31.73%
BenchmarkTDigestMerge/compression=1_n=1-4              2622          3454          +31.73%
BenchmarkTDigestMerge/compression=1_n=10-4             25603         22045         -13.90%
BenchmarkTDigestMerge/compression=1_n=100-4            213647        205863        -3.64%
BenchmarkTDigestMerge/compression=10_n=1-4             47663         30568         -35.87%
BenchmarkTDigestMerge/compression=10_n=10-4            244440        186289        -23.79%
BenchmarkTDigestMerge/compression=10_n=100-4           1742793       1655406       -5.01%
BenchmarkTDigestMerge/compression=20_n=1-4             129917        67920         -47.72%
BenchmarkTDigestMerge/compression=20_n=10-4            577894        394648        -31.71%
BenchmarkTDigestMerge/compression=20_n=100-4           3779385       3412943       -9.70%
BenchmarkTDigestMerge/compression=30_n=1-4             246840        100589        -59.25%
BenchmarkTDigestMerge/compression=30_n=10-4            1053926       632534        -39.98%
BenchmarkTDigestMerge/compression=30_n=100-4           5934629       5343870       -9.95%
BenchmarkTDigestMerge/compression=50_n=1-4             541190        191986        -64.53%
BenchmarkTDigestMerge/compression=50_n=10-4            2058839       1143808       -44.44%
BenchmarkTDigestMerge/compression=50_n=100-4           10637245      9538259       -10.33%
BenchmarkTDigestMerge/compression=100_n=1-4            1811738       466352        -74.26%
BenchmarkTDigestMerge/compression=100_n=10-4           6267957       2721893       -56.57%
BenchmarkTDigestMerge/compression=100_n=100-4          27109119      22270844      -17.85%

benchmark                                              old allocs     new allocs     delta
BenchmarkTDigestAddOnce/compression=1-4                0              0              +0.00%
BenchmarkTDigestAddOnce/compression=10-4               0              0              +0.00%
BenchmarkTDigestAddOnce/compression=20-4               0              0              +0.00%
BenchmarkTDigestAddOnce/compression=30-4               0              0              +0.00%
BenchmarkTDigestAddOnce/compression=50-4               0              0              +0.00%
BenchmarkTDigestAddOnce/compression=100-4              0              0              +0.00%
BenchmarkTDigestAddMulti/compression=1_n=10-4          7              5              -28.57%
BenchmarkTDigestAddMulti/compression=1_n=100-4         11             7              -36.36%
BenchmarkTDigestAddMulti/compression=1_n=1000-4        20             12             -40.00%
BenchmarkTDigestAddMulti/compression=1_n=10000-4       20             12             -40.00%
BenchmarkTDigestAddMulti/compression=10_n=10-4         7              5              -28.57%
BenchmarkTDigestAddMulti/compression=10_n=100-4        7              5              -28.57%
BenchmarkTDigestAddMulti/compression=10_n=1000-4       7              5              -28.57%
BenchmarkTDigestAddMulti/compression=10_n=10000-4      11             7              -36.36%
BenchmarkTDigestAddMulti/compression=20_n=10-4         7              5              -28.57%
BenchmarkTDigestAddMulti/compression=20_n=100-4        7              5              -28.57%
BenchmarkTDigestAddMulti/compression=20_n=1000-4       7              5              -28.57%
BenchmarkTDigestAddMulti/compression=20_n=10000-4      11             11             +0.00%
BenchmarkTDigestAddMulti/compression=30_n=10-4         7              5              -28.57%
BenchmarkTDigestAddMulti/compression=30_n=100-4        7              5              -28.57%
BenchmarkTDigestAddMulti/compression=30_n=1000-4       7              5              -28.57%
BenchmarkTDigestAddMulti/compression=30_n=10000-4      7              5              -28.57%
BenchmarkTDigestAddMulti/compression=50_n=10-4         7              5              -28.57%
BenchmarkTDigestAddMulti/compression=50_n=100-4        7              5              -28.57%
BenchmarkTDigestAddMulti/compression=50_n=1000-4       7              5              -28.57%
BenchmarkTDigestAddMulti/compression=50_n=10000-4      7              5              -28.57%
BenchmarkTDigestAddMulti/compression=100_n=10-4        7              5              -28.57%
BenchmarkTDigestAddMulti/compression=100_n=100-4       7              5              -28.57%
BenchmarkTDigestAddMulti/compression=100_n=1000-4      7              5              -28.57%
BenchmarkTDigestAddMulti/compression=100_n=10000-4     7              5              -28.57%
BenchmarkTDigestMerge/compression=1_n=1-4              14             10             -28.57%
BenchmarkTDigestMerge/compression=1_n=10-4             34             29             -14.71%
BenchmarkTDigestMerge/compression=1_n=100-4            218            214            -1.83%
BenchmarkTDigestMerge/compression=10_n=1-4             14             10             -28.57%
BenchmarkTDigestMerge/compression=10_n=10-4            32             28             -12.50%
BenchmarkTDigestMerge/compression=10_n=100-4           216            210            -2.78%
BenchmarkTDigestMerge/compression=20_n=1-4             14             10             -28.57%
BenchmarkTDigestMerge/compression=20_n=10-4            32             28             -12.50%
BenchmarkTDigestMerge/compression=20_n=100-4           216            214            -0.93%
BenchmarkTDigestMerge/compression=30_n=1-4             14             10             -28.57%
BenchmarkTDigestMerge/compression=30_n=10-4            32             28             -12.50%
BenchmarkTDigestMerge/compression=30_n=100-4           216            214            -0.93%
BenchmarkTDigestMerge/compression=50_n=1-4             14             10             -28.57%
BenchmarkTDigestMerge/compression=50_n=10-4            32             28             -12.50%
BenchmarkTDigestMerge/compression=50_n=100-4           216            214            -0.93%
BenchmarkTDigestMerge/compression=100_n=1-4            14             10             -28.57%
BenchmarkTDigestMerge/compression=100_n=10-4           32             28             -12.50%
BenchmarkTDigestMerge/compression=100_n=100-4          216            214            -0.93%

benchmark                                              old bytes     new bytes     delta
BenchmarkTDigestAddOnce/compression=1-4                0             0             +0.00%
BenchmarkTDigestAddOnce/compression=10-4               0             0             +0.00%
BenchmarkTDigestAddOnce/compression=20-4               0             0             +0.00%
BenchmarkTDigestAddOnce/compression=30-4               0             0             +0.00%
BenchmarkTDigestAddOnce/compression=50-4               0             0             +0.00%
BenchmarkTDigestAddOnce/compression=100-4              0             0             +0.00%
BenchmarkTDigestAddMulti/compression=1_n=10-4          336           256           -23.81%
BenchmarkTDigestAddMulti/compression=1_n=100-4         688           496           -27.91%
BenchmarkTDigestAddMulti/compression=1_n=1000-4        1824          1312          -28.07%
BenchmarkTDigestAddMulti/compression=1_n=10000-4       1880          1352          -28.09%
BenchmarkTDigestAddMulti/compression=10_n=10-4         1888          1440          -23.73%
BenchmarkTDigestAddMulti/compression=10_n=100-4        1888          1440          -23.73%
BenchmarkTDigestAddMulti/compression=10_n=1000-4       1888          1440          -23.73%
BenchmarkTDigestAddMulti/compression=10_n=10000-4      5504          4128          -25.00%
BenchmarkTDigestAddMulti/compression=20_n=10-4         3744          2816          -24.79%
BenchmarkTDigestAddMulti/compression=20_n=100-4        3744          2816          -24.79%
BenchmarkTDigestAddMulti/compression=20_n=1000-4       3744          2816          -24.79%
BenchmarkTDigestAddMulti/compression=20_n=10000-4      10560         10560         +0.00%
BenchmarkTDigestAddMulti/compression=30_n=10-4         5408          4096          -24.26%
BenchmarkTDigestAddMulti/compression=30_n=100-4        5408          4096          -24.26%
BenchmarkTDigestAddMulti/compression=30_n=1000-4       5408          4096          -24.26%
BenchmarkTDigestAddMulti/compression=30_n=10000-4      5408          4096          -24.26%
BenchmarkTDigestAddMulti/compression=50_n=10-4         8352          6272          -24.90%
BenchmarkTDigestAddMulti/compression=50_n=100-4        8352          6272          -24.90%
BenchmarkTDigestAddMulti/compression=50_n=1000-4       8352          6272          -24.90%
BenchmarkTDigestAddMulti/compression=50_n=10000-4      8352          6272          -24.90%
BenchmarkTDigestAddMulti/compression=100_n=10-4        16544         12416         -24.95%
BenchmarkTDigestAddMulti/compression=100_n=100-4       16544         12416         -24.95%
BenchmarkTDigestAddMulti/compression=100_n=1000-4      16544         12416         -24.95%
BenchmarkTDigestAddMulti/compression=100_n=10000-4     16544         12416         -24.95%
BenchmarkTDigestMerge/compression=1_n=1-4              590           535           -9.32%
BenchmarkTDigestMerge/compression=1_n=10-4             1987          1789          -9.96%
BenchmarkTDigestMerge/compression=1_n=100-4            12349         12760         +3.33%
BenchmarkTDigestMerge/compression=10_n=1-4             3486          2824          -18.99%
BenchmarkTDigestMerge/compression=10_n=10-4            10678         9804          -8.19%
BenchmarkTDigestMerge/compression=10_n=100-4           78583         77992         -0.75%
BenchmarkTDigestMerge/compression=20_n=1-4             6934          5568          -19.70%
BenchmarkTDigestMerge/compression=20_n=10-4            21368         19429         -9.07%
BenchmarkTDigestMerge/compression=20_n=100-4           158496        154471        -2.54%
BenchmarkTDigestMerge/compression=30_n=1-4             9952          7807          -21.55%
BenchmarkTDigestMerge/compression=30_n=10-4            31546         28464         -9.77%
BenchmarkTDigestMerge/compression=30_n=100-4           232436        230332        -0.91%
BenchmarkTDigestMerge/compression=50_n=1-4             15948         12697         -20.39%
BenchmarkTDigestMerge/compression=50_n=10-4            51043         46656         -8.59%
BenchmarkTDigestMerge/compression=50_n=100-4           385184        381185        -1.04%
BenchmarkTDigestMerge/compression=100_n=1-4            31937         25775         -19.29%
BenchmarkTDigestMerge/compression=100_n=10-4           102726        94753         -7.76%
BenchmarkTDigestMerge/compression=100_n=100-4          785440        781312        -0.53%
```

That results in significant improvement but just removing fenwick tree is even better. Following benchcmp is fenwick_threshold vs fenwick_disabled:

```
benchmark                                              old ns/op     new ns/op     delta
BenchmarkTDigestAddOnce/compression=1-4                156           157           +0.64%
BenchmarkTDigestAddOnce/compression=10-4               210           210           +0.00%
BenchmarkTDigestAddOnce/compression=20-4               218           249           +14.22%
BenchmarkTDigestAddOnce/compression=30-4               229           276           +20.52%
BenchmarkTDigestAddOnce/compression=50-4               241           324           +34.44%
BenchmarkTDigestAddOnce/compression=100-4              260           427           +64.23%
BenchmarkTDigestAddMulti/compression=1_n=10-4          1561          1501          -3.84%
BenchmarkTDigestAddMulti/compression=1_n=100-4         14633         14875         +1.65%
BenchmarkTDigestAddMulti/compression=1_n=1000-4        165089        161146        -2.39%
BenchmarkTDigestAddMulti/compression=1_n=10000-4       1536685       1540990       +0.28%
BenchmarkTDigestAddMulti/compression=10_n=10-4         2228          2252          +1.08%
BenchmarkTDigestAddMulti/compression=10_n=100-4        16744         16437         -1.83%
BenchmarkTDigestAddMulti/compression=10_n=1000-4       200966        203042        +1.03%
BenchmarkTDigestAddMulti/compression=10_n=10000-4      2062918       2080249       +0.84%
BenchmarkTDigestAddMulti/compression=20_n=10-4         2646          2642          -0.15%
BenchmarkTDigestAddMulti/compression=20_n=100-4        20279         18964         -6.48%
BenchmarkTDigestAddMulti/compression=20_n=1000-4       224492        226276        +0.79%
BenchmarkTDigestAddMulti/compression=20_n=10000-4      2288713       2297733       +0.39%
BenchmarkTDigestAddMulti/compression=30_n=10-4         3059          3042          -0.56%
BenchmarkTDigestAddMulti/compression=30_n=100-4        22113         21266         -3.83%
BenchmarkTDigestAddMulti/compression=30_n=1000-4       245660        240281        -2.19%
BenchmarkTDigestAddMulti/compression=30_n=10000-4      2420602       2424417       +0.16%
BenchmarkTDigestAddMulti/compression=50_n=10-4         3606          3560          -1.28%
BenchmarkTDigestAddMulti/compression=50_n=100-4        25500         24514         -3.87%
BenchmarkTDigestAddMulti/compression=50_n=1000-4       272792        273560        +0.28%
BenchmarkTDigestAddMulti/compression=50_n=10000-4      2732537       2710829       -0.79%
BenchmarkTDigestAddMulti/compression=100_n=10-4        4858          4815          -0.89%
BenchmarkTDigestAddMulti/compression=100_n=100-4       31295         29820         -4.71%
BenchmarkTDigestAddMulti/compression=100_n=1000-4      347337        339679        -2.20%
BenchmarkTDigestAddMulti/compression=100_n=10000-4     3341165       3317960       -0.69%
BenchmarkTDigestMerge/compression=1_n=1-4              3454          2337          -32.34%
BenchmarkTDigestMerge/compression=1_n=10-4             22045         20810         -5.60%
BenchmarkTDigestMerge/compression=1_n=100-4            205863        199081        -3.29%
BenchmarkTDigestMerge/compression=10_n=1-4             30568         28018         -8.34%
BenchmarkTDigestMerge/compression=10_n=10-4            186289        181264        -2.70%
BenchmarkTDigestMerge/compression=10_n=100-4           1655406       1633496       -1.32%
BenchmarkTDigestMerge/compression=20_n=1-4             67920         67638         -0.42%
BenchmarkTDigestMerge/compression=20_n=10-4            394648        396629        +0.50%
BenchmarkTDigestMerge/compression=20_n=100-4           3412943       3499240       +2.53%
BenchmarkTDigestMerge/compression=30_n=1-4             100589        105376        +4.76%
BenchmarkTDigestMerge/compression=30_n=10-4            632534        616610        -2.52%
BenchmarkTDigestMerge/compression=30_n=100-4           5343870       5445699       +1.91%
BenchmarkTDigestMerge/compression=50_n=1-4             191986        191644        -0.18%
BenchmarkTDigestMerge/compression=50_n=10-4            1143808       1126963       -1.47%
BenchmarkTDigestMerge/compression=50_n=100-4           9538259       10102338      +5.91%
BenchmarkTDigestMerge/compression=100_n=1-4            466352        461295        -1.08%
BenchmarkTDigestMerge/compression=100_n=10-4           2721893       2694022       -1.02%
BenchmarkTDigestMerge/compression=100_n=100-4          22270844      24744200      +11.11%

benchmark                                              old allocs     new allocs     delta
BenchmarkTDigestAddOnce/compression=1-4                0              0              +0.00%
BenchmarkTDigestAddOnce/compression=10-4               0              0              +0.00%
BenchmarkTDigestAddOnce/compression=20-4               0              0              +0.00%
BenchmarkTDigestAddOnce/compression=30-4               0              0              +0.00%
BenchmarkTDigestAddOnce/compression=50-4               0              0              +0.00%
BenchmarkTDigestAddOnce/compression=100-4              0              0              +0.00%
BenchmarkTDigestAddMulti/compression=1_n=10-4          5              5              +0.00%
BenchmarkTDigestAddMulti/compression=1_n=100-4         7              7              +0.00%
BenchmarkTDigestAddMulti/compression=1_n=1000-4        12             7              -41.67%
BenchmarkTDigestAddMulti/compression=1_n=10000-4       12             12             +0.00%
BenchmarkTDigestAddMulti/compression=10_n=10-4         5              5              +0.00%
BenchmarkTDigestAddMulti/compression=10_n=100-4        5              5              +0.00%
BenchmarkTDigestAddMulti/compression=10_n=1000-4       5              5              +0.00%
BenchmarkTDigestAddMulti/compression=10_n=10000-4      7              7              +0.00%
BenchmarkTDigestAddMulti/compression=20_n=10-4         5              5              +0.00%
BenchmarkTDigestAddMulti/compression=20_n=100-4        5              5              +0.00%
…