@caio 2018-11-08
Hot damn! This goes to show that I shouldn't trust micro benchmarks as representative of the whole thing :sweat_smile: 

Thanks a lot for all the work, especially for adding more realistic benchmarks! The improvements would be even more noticeable before your initial stab at making Add reuse the tree when possible and avoiding the extra allocation on rebuild.

I've been holding out on releasing a new version of the package because it looks like you're on a roll :grin: I'll release a new one this coming weekend but- as before- if you need an earlier release just let me know!